Split `bound.rs` into small files
